Split the simulator up.
